refactor(platforms): huya api mode configuration
