+++ crates/platforms/src/extractor/platforms/huya/mod.rs
@@ -2,11 +2,17 @@ mod builder;
 pub mod danmu;
 pub(crate) mod danmu_models;
 mod models;
+mod mp;
+mod sign;
 pub mod tars;
+mod web;
+mod wup;
 
 pub use builder::Huya;
 pub use builder::URL_REGEX;
 pub use tars::*;
+
+pub use sign::get_anticode;
 
 // Re-export the danmu provider
 pub use danmu::HuyaDanmuProvider;
